LayoutTests/media/track/track-cue-rendering-inner-timestamps.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script src=../media-controls.js></script>

        <script>

        var testTrack;
        var cueDisplayElement;
        var cueDisplayContainer;
        var seekedCount = 0;

        var testCaseDescription = [
                "First cue (empty)",
                "First cue (one word)",
                "First cue (two words)",
                "First cue (three words)",
                "Second cue (empty)",
                "Second cue (one word)",
                "Second cue (two words)",
                "Second cue (three words)",
                "Second cue (same as before)",
                "Second cue (four words)",
                "Second cue (five words)",
                "Second cue (same as before)",
                "Second cue (six words)",
                "Second cue (seven words)",
                "Second cue (eight words)",
                "Third cue (one word)",
                "Third cue (two words)"
        ];

        // Each test case is found by incrementing with 0.250 the current time.
        var testCaseSeekTime = [
        ];

        var currentTestCase = 0;

        function seeked()
        {
            for (var i = 0; i < testTrack.track.activeCues.length; ++i) {
                cueDisplayElement = textTrackDisplayElement(video, 'display', i);

                consoleWrite("");
                consoleWrite("* Cue " + i + " *");

                consoleWrite("Current time: " + Math.round(video.currentTime * 100) / 100);
                consoleWrite("Past: " + cueDisplayElement.childNodes[0].innerHTML);
                consoleWrite("Future: " + cueDisplayElement.childNodes[1].innerHTML);

                consoleWrite("Complete: " + testTrack.track.activeCues[i].text);
            }

            runNextTestCase();
        };

        function runNextTestCase()
        {
            if (currentTestCase == testCaseDescription.length) {
                endTest();
                return;
            }

            consoleWrite("");
            consoleWrite("** " +  testCaseDescription[currentTestCase] + " **");

            video.currentTime = testCaseSeekTime[currentTestCase];
            waitForEvent('seeked', seeked, false, true);

            currentTestCase++;
        }

        function loaded()
        {
            consoleWrite("Test that TextTrack's cues are rendered correctly when they contain inner timestamps.");

            findMediaElement();
            testTrack = document.querySelector('track');
            video.src = findMediaFile('video', '../content/test');

            waitForEvent('canplaythrough', function() {
                cueDisplayContainer = textTrackDisplayElement(video);

                for (i = 0.0; i < 4.500; i += 0.250)
                    testCaseSeekTime.push(i);

                runNextTestCase();
            });
        }

        </script>
    </head>
    <body onload="loaded()">
        <video controls>
            <track src="captions-webvtt/captions-inner-timestamps.vtt" kind="captions" default>
        </video>
    </body>
</html>
