LayoutTests/media/track/track-cue-rendering-inner-timestamps-expected.txt

Test that TextTrack's cues are rendered correctly when they contain inner timestamps.
EVENT(canplaythrough)

** First cue (empty) **
EVENT(seeked)

* Cue 0 *
Current time: 0
Past: 
Future: Chocolate Rain Karaoke
Complete: <00:00:00.200>Chocolate <00:00:00.450>Rain <00:00:00.700>Karaoke

** First cue (one word) **
EVENT(seeked)

* Cue 0 *
Current time: 0.25
Past: Chocolate 
Future: Rain Karaoke
Complete: <00:00:00.200>Chocolate <00:00:00.450>Rain <00:00:00.700>Karaoke

** First cue (two words) **
EVENT(seeked)

* Cue 0 *
Current time: 0.5
Past: Chocolate Rain 
Future: Karaoke
Complete: <00:00:00.200>Chocolate <00:00:00.450>Rain <00:00:00.700>Karaoke

** First cue (three words) **
EVENT(seeked)

* Cue 0 *
Current time: 0.75
Past: Chocolate Rain Karaoke
Future: 
Complete: <00:00:00.200>Chocolate <00:00:00.450>Rain <00:00:00.700>Karaoke

** Second cue (empty) **
EVENT(seeked)

* Cue 0 *
Current time: 1
Past: 
Future: Some stay dry and others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (one word) **
EVENT(seeked)

* Cue 0 *
Current time: 1.25
Past: Some 
Future: stay dry and others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (two words) **
EVENT(seeked)

* Cue 0 *
Current time: 1.5
Past: Some stay 
Future: dry and others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (three words) **
EVENT(seeked)

* Cue 0 *
Current time: 1.75
Past: Some stay dry 
Future: and others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (same as before) **
EVENT(seeked)

* Cue 0 *
Current time: 2
Past: Some stay dry 
Future: and others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (four words) **
EVENT(seeked)

* Cue 0 *
Current time: 2.25
Past: Some stay dry and 
Future: others feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (five words) **
EVENT(seeked)

* Cue 0 *
Current time: 2.5
Past: Some stay dry and others 
Future: feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (same as before) **
EVENT(seeked)

* Cue 0 *
Current time: 2.75
Past: Some stay dry and others 
Future: feel the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (six words) **
EVENT(seeked)

* Cue 0 *
Current time: 3
Past: Some stay dry and others feel 
Future: the pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (seven words) **
EVENT(seeked)

* Cue 0 *
Current time: 3.25
Past: Some stay dry and others feel the 
Future: pain
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Second cue (eight words) **
EVENT(seeked)

* Cue 0 *
Current time: 3.5
Past: Some stay dry and others feel the pain
Future: 
Complete: <00:00:01.200>Some <00:00:01.450>stay <00:00:01.700>dry <00:00:02.200>and <00:00:02.450>others <00:00:02.950>feel <00:00:03.200>the <00:00:03.450>pain

** Third cue (one word) **
EVENT(seeked)

* Cue 0 *
Current time: 3.75
Past: Chocolate 
Future: Rain
Complete: <00:00:03.750>Chocolate <00:00:04.000>Rain

** Third cue (two words) **
EVENT(seeked)

* Cue 0 *
Current time: 4
Past: Chocolate Rain
Future: 
Complete: <00:00:03.750>Chocolate <00:00:04.000>Rain
END OF TEST
